fix: centralize game finalization for stale AI games
